backend/requirements.txt
matplotlib
numpy
tensorflow==1.15.0
keras==2.2.5
flask
opencv-python
flask-cors
Pillow
